requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bidict==0.23.1
click==8.3.1
fastapi==0.124.0
h11==0.16.0
idna==3.11
netifaces==0.10.6
numpy==2.3.5
pydantic==2.12.5
pydantic_core==2.41.5
python-engineio==4.12.3
python-socketio==5.15.0
setuptools==3.3
simple-websocket==1.1.0
SQLAlchemy==2.0.44
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
websockets==15.0.1
wsproto==1.3.2
